breaks_timespan {scales}	R Documentation

Breaks for timespan data

Description

As timespan units span a variety of bases (1000 below seconds, 60 for second and minutes, 24 for hours, and 7 for days), the range of the input data determines the base used for calculating breaks

Usage

breaks_timespan(unit = c("secs", "mins", "hours", "days", "weeks"), n = 5)

Arguments

unit
The unit used to interpret numeric data input

n
Desired number of breaks. You may get slightly more or fewer breaks that requested.

Examples

demo_timespan(seq(0, 100), breaks = breaks_timespan())

[Package scales version 1.3.0 Index]