Formatted dates and times.

date_short() attempts to construct the shortest format string that will uniquely identify dates. It's inspired by matplotlib's [ConciseDateFormatter``](https://matplotlib.org/api/dates_api.html#matplotlib.dates.ConciseDateFormatter), but uses a slightly different approach: ConciseDateFormatterformats "firsts" (e.g. first day of month, first day of day) specially;date_short()` formats changes (e.g. new month, new year) specially.

date_format(format = "%Y-%m-%d", tz = "UTC")

date_short(format = c("%Y", "%b", "%d", "%H:%M"), sep = "\n")

time_format(format = "%H:%M:%S", tz = "UTC")

Arguments

format
For date_format() and time_format() a date/time format string using standard POSIX specification. See strptime() for details.

For date_short() a character vector of length 4 giving the format components to use for year, month, day, and hour respectively.

tz
a time zone name, see timezones(). Defaults to UTC

sep
Separator to use when combining date formats into a single string.

Examples

a_time <- ISOdatetime(2012, 1, 1, 11, 30, 0, tz = "UTC")
a_date <- as.Date(a_time)

date_format()(a_date)

#> [1] "2012-01-01"

date_format(format = "%A")(a_date)

#> [1] "Sunday"

x <- as.Date("2012-12-25") + (0:20) * 2
date_short()(x)

#>  [1] "25\nDec\n2012" "27"            "29"            "31"           
#>  [5] "02\nJan\n2013" "04"            "06"            "08"           
#>  [9] "10"            "12"            "14"            "16"           
#> [13] "18"            "20"            "22"            "24"           
#> [17] "26"            "28"            "30"            "01\nFeb"      
#> [21] "03"           

x <- as.POSIXct("2012-12-25") + (0:10) * 3600 * 10
date_short()(x)

#>  [1] "00:00\n25\nDec\n2012" "10:00"                "20:00"               
#>  [4] "06:00\n26"            "16:00"                "02:00\n27"           
#>  [7] "12:00"                "22:00"                "08:00\n28"           
#> [10] "18:00"                "04:00\n29"           

time_format()(a_time)

#> [1] "11:30:00"

time_format(tz = "Europe/Berlin")(a_time)

#> [1] "12:30:00"

a_hms <- hms::as.hms(a_time, tz = "UTC")

#> Warning: as.hms() is deprecated, please use as_hms().
#> This warning is displayed once per session.

time_format(format = "%H:%M")(a_hms)

#> [1] "11:30"