expand_range {scales}	R Documentation

Expand a range with a multiplicative or additive constant.

Description

Expand a range with a multiplicative or additive constant.

Usage

  expand_range(range, mul = 0, add = 0, zero_width = 1)

Arguments

range
range of data, numeric vector of length 2

mul
multiplicative constract

add
additive constant

zero_width
distance to use if range has zero width

[Package scales version 0.2.3 Index]