extended_breaks {scales}	R Documentation

Extended breaks. Uses Wilkinson's extended breaks algorithm as implemented in the labeling package.

Description

Extended breaks. Uses Wilkinson's extended breaks algorithm as implemented in the labeling package.

Usage

  extended_breaks(n = 5, ...)

Arguments

n
desired number of breaks

...
other arguments passed on to extended

References

Talbot, J., Lin, S., Hanrahan, P. (2010) An Extension of Wilkinson's Algorithm for Positioning Tick Labels on Axes, InfoVis 2010.

Examples

extended_breaks()(1:10)
extended_breaks()(1:100)

[Package scales version 0.2.3 Index]