gradient_n_pal {scales}	R Documentation

Arbitrary colour gradient palette (continous).

Description

Arbitrary colour gradient palette (continous).

Usage

  gradient_n_pal(colours, values = NULL, space = "Lab")

Arguments

colours
vector of colours

values
if colours should not be evenly positioned along the gradient this vector gives the position (between 0 and 1) for each colour in the colours vector. See rescale for a convience function to map an arbitrary range to between 0 and 1.

space
colour space in which to calculate gradient. "Lab" usually best unless gradient goes through white.

[Package scales version 0.2.3 Index]