parse_format {scales}	R Documentation

Parse a text label to produce expressions for plotmath.

Description

Parse a text label to produce expressions for plotmath.

Usage

parse_format()

Value

a function with single parameter x, a character vector, that returns a list of expressions

See Also

plotmath()

Examples

parse_format()(c("alpha", "beta", "gamma"))

[Package scales version 1.0.0 Index]