pretty_breaks {scales}	R Documentation

Pretty breaks. Uses default R break algorithm as implemented in pretty.

Description

Pretty breaks. Uses default R break algorithm as implemented in pretty.

Usage

  pretty_breaks(n = 5, ...)

Arguments

n
desired number of breaks

...
other arguments passed on to pretty

Examples

pretty_breaks()(1:10)
pretty_breaks()(1:100)
pretty_breaks()(as.Date(c("2008-01-01", "2009-01-01")))
pretty_breaks()(as.Date(c("2008-01-01", "2090-01-01")))

[Package scales version 0.2.3 Index]