pseudo_log_trans {scales}	R Documentation

Pseudo-log transformation

Description

A transformation mapping numbers to a signed logarithmic scale with a smooth transition to linear scale around 0.

Usage

pseudo_log_trans(sigma = 1, base = exp(1))

Arguments

sigma
scaling factor for the linear part

base
approximate logarithm base used

[Package scales version 1.0.0 Index]