rescale {scales}	R Documentation

Rescale numeric vector to have specified minimum and maximum.

Description

Rescale numeric vector to have specified minimum and maximum.

Usage

  rescale(x, to = c(0, 1), from = range(x, na.rm = TRUE))

Arguments

x
numeric vector of values to manipulate.

to
output range (numeric vector of length two)

from
input range (numeric vector of length two). If not given, is calculated from the range of x

Examples

rescale(1:100)
rescale(runif(50))
rescale(1)

[Package scales version 0.2.3 Index]