rescale_mid {scales}	R Documentation

Rescale numeric vector to have specified minimum, midpoint, and maximum.

Description

Rescale numeric vector to have specified minimum, midpoint, and maximum.

Usage

  rescale_mid(x, to = c(0, 1),
    from = range(x, na.rm = TRUE), mid = 0)

Arguments

x
numeric vector of values to manipulate.

to
output range (numeric vector of length two)

from
input range (numeric vector of length two). If not given, is calculated from the range of x

mid
mid-point of input range

Examples

rescale_mid(1:100, mid = 50.5)
rescale_mid(runif(50), mid = 0.5)
rescale_mid(1)

[Package scales version 0.2.3 Index]