trans_format {scales}	R Documentation

Format labels after transformation

Description

Usage

trans_format(trans, format = scientific_format())

Arguments

trans
transformation to apply

format
additional formatter to apply after transformation

Value

a function with single parameter x, a numeric vector, that returns a character vector of list of expressions

Examples

tf <- trans_format("log10", scientific_format())
tf(10^1:6)

[Package scales version 1.2.1 Index]