trans_range {scales}	R Documentation

Compute range of transformed values

Description

Silently drops any ranges outside of the domain of trans.

Usage

trans_range(trans, x)

Arguments

trans
a transformation object, or the name of a transformation object given as a string.

x
a numeric vector to compute the range of

[Package scales version 1.2.1 Index]