label_log {scales}	R Documentation

Label numbers in log format (10^3, 10^6, etc)

Description

label_log() displays numbers as base^exponent, using superscript formatting.

Usage

label_log(base = 10, digits = 3)

Arguments

base
Base of logarithm to use

digits
Number of significant digits to show for the exponent. Argument is passed on to base::format().

Value

All label_() functions return a "labelling" function, i.e. a function that takes a vector x and returns a character vector of length(x) giving a label for each input value.

Labelling functions are designed to be used with the labels argument of ggplot2 scales. The examples demonstrate their use with x scales, but they work similarly for all scales, including those that generate legends rather than axes.

See Also

breaks_log() for the related breaks algorithm.

Other labels for log scales: label_bytes(), label_number_si(), label_scientific()

Examples

demo_log10(c(1, 1e5), labels = label_log())
demo_log10(c(1, 1e5), breaks = breaks_log(base = 2), labels = label_log(base = 2))

[Package scales version 1.2.0 Index]