math_format {scales}	R Documentation

Add arbitrary expression to a label. The symbol that will be replace by the label value is .x.

Description

Add arbitrary expression to a label. The symbol that will be replace by the label value is .x.

Usage

math_format(expr = 10^.x, format = force)

Arguments

expr
expression to use

format
another format function to apply prior to mathematical transformation - this makes it easier to use floating point numbers in mathematical expressions.

Value

a function with single parameter x, a numeric vector, that returns a list of expressions

See Also

plotmath()

Examples

math_format()(1:10)
math_format(alpha + frac(1, .x))(1:10)
math_format()(runif(10))
math_format(format = percent)(runif(10))

[Package scales version 1.0.0 Index]