parse_format {scales}	R Documentation

Superseded interface to label_parse()/label_math()

Description

These functions are kept for backward compatibility; you should switch to label_parse()/label_math() for new code.

Usage

parse_format()

math_format(expr = 10^.x, format = force)

Arguments

expr
expression to use

format
another format function to apply prior to mathematical transformation - this makes it easier to use floating point numbers in mathematical expressions.

[Package scales version 1.2.0 Index]