rescale_pal {scales}	R Documentation

Rescale palette (continuous)

Description

Just rescales the input to the specific output range. Useful for alpha, size, and continuous position.

Usage

rescale_pal(range = c(0.1, 1))

Arguments

range
Numeric vector of length two, giving range of possible values. Should be between 0 and 1.

[Package scales version 1.2.0 Index]