scales-package {scales}	R Documentation

scales: Scale Functions for Visualization

Description

Graphical scales map data to aesthetics, and provide methods for automatically determining breaks and labels for axes and legends.

Author(s)

Maintainer: Hadley Wickham hadley@rstudio.com

Authors:

Dana Seidel

Other contributors:

RStudio [copyright holder, funder]

See Also

Useful links:

https://scales.r-lib.org

https://github.com/r-lib/scales

Report bugs at https://github.com/r-lib/scales/issues

[Package scales version 1.2.0 Index]