scientific_format {scales}	R Documentation

Superseded interface to label_scientific()

Description

These functions are kept for backward compatibility; you should switch to label_scientific() for new code.

Usage

scientific_format(
  digits = 3,
  scale = 1,
  prefix = "",
  suffix = "",
  decimal.mark = ".",
  trim = TRUE,
  ...
)

scientific(
  x,
  digits = 3,
  scale = 1,
  prefix = "",
  suffix = "",
  decimal.mark = ".",
  trim = TRUE,
  ...
)

Arguments

digits
Number of digits to show before exponent.

scale
A scaling factor: x will be multiplied by scale before formatting. This is useful if the underlying data is very small or very large.

prefix
Symbols to display before and after value.

suffix
Symbols to display before and after value.

decimal.mark
The character to be used to indicate the numeric decimal point.

trim
Logical, if FALSE, values are right-justified to a common width (see base::format()).

...
Other arguments passed on to base::format().

[Package scales version 1.2.0 Index]