yj_trans {scales}	R Documentation

Yeo-Johnson transformation

Description

The Yeo-Johnson transformation is a flexible transformation that is similiar to Box-Cox, boxcox_trans(), but does not require input values to be greater than zero.

Usage

yj_trans(p)

Arguments

p
Transformation exponent, λ.

Details

The transformation takes one of four forms depending on the values of y and λ.

y ≥ 0 and λ != 0 : y^(λ) = ((y + 1)^λ - 1)/λ

y ≥ 0 and λ = 0: y^(λ) = ln(y + 1)

y < 0 and λ != 2: y^(λ) = -((-y + 1)^(2 - λ) - 1)/(2 - λ)

y < 0 and λ = 2: y^(λ) = -ln(-y + 1)

References

Yeo, I., & Johnson, R. (2000). A New Family of Power Transformations to Improve Normality or Symmetry. Biometrika, 87(4), 954-959. http://www.jstor.org/stable/2673623

Examples

plot(yj_trans(-1), xlim = c(-10, 10))
plot(yj_trans(0), xlim = c(-10, 10))
plot(yj_trans(1), xlim = c(-10, 10))
plot(yj_trans(2), xlim = c(-10, 10))

[Package scales version 1.2.0 Index]